requirements.txt
colorama==0.3.9
nltk==3.2.5
numpy==1.14.0
six==1.11.0
tensorflow-gpu
tqdm==4.19.5
sklearn
seaborn
matplotlib
